[refactor] separate cli program per module
